+++ mite_web/mite_web/prep_data/download_manager.py
@@ -37,18 +37,14 @@ class DownloadManager(BaseModel):
     """Download data and prepare for use by mite_web
 
     Attributes:
-        record_url: Zenodo URL for mite_data: always resolves to latest version
+        mite_data: Zenodo URL for mite_data: always resolves to latest version
+        mite_web_extras: Zenodo URL for mite_web_extras: always resolves to latest version
         location: the location to download data to
-        record: path to the record file
-        record_unzip: path to unzipped record file
-        version: path to the mite_data version location
     """
 
-    record_url: str = "https://zenodo.org/api/records/13294303"
+    mite_data: str = "https://zenodo.org/api/records/13294303"
+    mite_web_extras: str = "https://zenodo.org/api/records/17453501"
     location: Path = Path(__file__).parent.parent.joinpath("data")
-    record: Path = Path(__file__).parent.parent.joinpath("data/record.zip")
-    record_unzip: Path = Path(__file__).parent.parent.joinpath("data/record")
-    version: Path = Path(__file__).parent.parent.joinpath("data/version.json")
 
     def run(self) -> None:
         """Call methods for downloading and moving data"""
@@ -60,56 +56,96 @@ class DownloadManager(BaseModel):
                 f"DownloadManager: Download location {self.location} already exists - skip download"
             )
             return
+        self.location.mkdir()
 
-        self.location.mkdir(parents=True)
-        self.download_data()
-        self.organize_data()
+        self.download_mite_data()
+        self.organize_mite_data()
+
+        self.download_mite_web_extras()
+        self.organize_mite_web_extras()
+
+        self.generate_summary()
 
         logger.info("DownloadManager: Completed")
 
-    def download_data(self) -> None:
-        """Download data from Zenodo
+    def download_mite_data(self) -> None:
+        """Download and store mite_data record from Zenodo
 
         Raises:
             RuntimeError: Could not download files
         """
-        response_metadata = requests.get(self.record_url)
-        if response_metadata.status_code != 200:
-            raise RuntimeError(
-                f"Error fetching 'mite_data' record metadata: {response_metadata.status_code}"
-            )
+        logger.info("DownloadManager: Start mite_data download")
 
-        record_metadata = response_metadata.json()
+        rsps_meta = requests.get(self.mite_data)
+        if rsps_meta.status_code != 200:
+            raise RuntimeError(
+                f"Error fetching 'mite_data' metadata: {rsps_meta.status_code}"
+            )
+        record_metadata = rsps_meta.json()
         version = record_metadata["metadata"]["version"]
         files_url = record_metadata["files"][0]["links"]["self"]
 
-        response_data = requests.get(files_url)
-        if response_data.status_code != 200:
+        rsps_data = requests.get(files_url)
+        if rsps_data.status_code != 200:
             raise RuntimeError(
-                f"Error downloading 'mite_data' record: {response_data.status_code}"
+                f"Error downloading 'mite_data' record: {rsps_data.status_code}"
             )
 
-        with open(self.record, "wb") as f:
-            f.write(response_data.content)
-        with open(self.version, "w") as f:
+        with open(self.location.joinpath("mite_data.zip"), "wb") as f:
+            f.write(rsps_data.content)
+
+        with open(self.location.joinpath("version.json"), "w") as f:
             f.write(json.dumps({"version_mite_data": f"{version}"}))
 
-    def organize_data(self) -> None:
-        """Unpacks data, moves to convenient location, cleans up
+        logger.info("DownloadManager: Completed mite_data download")
+
+    def download_mite_web_extras(self) -> None:
+        """Download and store mite_web_extras record from Zenodo
+
+        Raises:
+            RuntimeError: Could not download files
+        """
+        logger.info("DownloadManager: Started mite_web_extras download")
+
+        rsps_meta = requests.get(self.mite_web_extras)
+        if rsps_meta.status_code != 200:
+            raise RuntimeError(
+                f"Error fetching 'mite_web_extras' metadata: {rsps_meta.status_code}"
+            )
+        record_metadata = rsps_meta.json()
+        files_url = record_metadata["files"][0]["links"]["self"]
+
+        rsps_data = requests.get(files_url)
+        if rsps_data.status_code != 200:
+            raise RuntimeError(
+                f"Error downloading 'mite_web_extras' record: {rsps_data.status_code}"
+            )
+
+        with open(self.location.joinpath("mite_web_extras.zip"), "wb") as f:
+            f.write(rsps_data.content)
+
+        logger.info("DownloadManager: Completed mite_web_extras download")
+
+    def organize_mite_data(self) -> None:
+        """Unpacks mite_data, moves to correct locations, cleans up
 
         Raises:
             NotADirectoryError: directory not unzipped in expected location
             RuntimeError: Could not determine data location in downloaded folder
         """
-        shutil.unpack_archive(
-            filename=self.record, extract_dir=self.record_unzip, format="zip"
-        )
-        if not self.record_unzip.exists():
-            raise NotADirectoryError(
-                f"Could not find the unzipped directory {self.record_unzip}."
-            )
+        logger.info("DownloadManager: Started to organize mite_data download")
 
-        matching_dirs = list(self.record_unzip.glob("mite-standard-mite_data-*"))
+        src = self.location.joinpath("mite_data.zip")
+        trgt = self.location.joinpath("mite_data_unzip")
+        download = self.location.joinpath("download")
+
+        download.mkdir(exist_ok=True)
+
+        shutil.unpack_archive(filename=src, extract_dir=trgt, format="zip")
+        if not trgt.exists():
+            raise NotADirectoryError(f"Could not find the unzipped directory {trgt}.")
+
+        matching_dirs = list(trgt.glob("mite-standard-mite_data-*"))
         if not matching_dirs:
             raise RuntimeError(
                 f"Could not determine data storage location in downloaded directory."
@@ -117,16 +153,158 @@ class DownloadManager(BaseModel):
         subdir = matching_dirs[0]
 
         shutil.move(
-            src=self.record_unzip.joinpath(subdir).joinpath("mite_data/data").resolve(),
+            src=trgt.joinpath(subdir).joinpath("mite_data/data").resolve(),
             dst=self.location.resolve(),
         )
-
         shutil.move(
-            src=self.record_unzip.joinpath(subdir)
-            .joinpath("mite_data/fasta")
+            src=trgt.joinpath(subdir).joinpath("mite_data/fasta").resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/dump_smarts.csv")
+            .resolve(),
+            dst=download.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/dump_smiles.csv")
+            .resolve(),
+            dst=download.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/metadata_general.json")
+            .resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/metadata_mibig.json")
+            .resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/product_list.pickle")
+            .resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/reaction_fps.pickle")
+            .resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/substrate_list.pickle")
+            .resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir)
+            .joinpath("mite_data/metadata/summary.csv")
             .resolve(),
             dst=self.location.resolve(),
         )
 
-        os.remove(self.record)
-        shutil.rmtree(self.record_unzip)
+        os.remove(src)
+        shutil.rmtree(trgt)
+
+        logger.info("DownloadManager: Completed to organize mite_data download")
+
+    def organize_mite_web_extras(self) -> None:
+        """Unpacks mite_web_extras, moves to correct locations, cleans up
+
+        Raises:
+            NotADirectoryError: directory not unzipped in expected location
+            RuntimeError: Could not determine data location in downloaded folder
+        """
+        logger.info("DownloadManager: Started to organize mite_web_extras download")
+
+        src = self.location.joinpath("mite_web_extras.zip")
+        trgt = self.location.joinpath("mite_web_extras_unzip")
+        download = self.location.joinpath("download")
+
+        download.mkdir(exist_ok=True)
+
+        shutil.unpack_archive(filename=src, extract_dir=trgt, format="zip")
+        if not trgt.exists():
+            raise NotADirectoryError(f"Could not find the unzipped directory {trgt}.")
+
+        matching_dirs = list(trgt.glob("mite-standard-mite_web_extras-*"))
+        if not matching_dirs:
+            raise RuntimeError(
+                f"Could not determine data storage location in downloaded directory."
+            )
+        subdir = matching_dirs[0]
+
+        shutil.move(
+            src=trgt.joinpath(subdir).joinpath("data/blastlib").resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir).joinpath("data/html").resolve(),
+            dst=self.location.resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir).joinpath("data/img").resolve(),
+            dst=self.location.parent.joinpath("static").resolve(),
+        )
+        shutil.move(
+            src=trgt.joinpath(subdir).joinpath("data/mite_concat.fasta").resolve(),
+            dst=download.resolve(),
+        )
+
+        with open(self.location.joinpath("version.json")) as infile:
+            v_mite_data = json.load(infile)
+
+        with open(trgt.joinpath(subdir).joinpath("data/version.json")) as infile:
+            v_mite_web_extr = json.load(infile)
+
+        if v_mite_data["version_mite_data"] != v_mite_web_extr["version_mite_data"]:
+            raise RuntimeError(
+                f"Version mite_data {v_mite_data["version_mite_data"]} does not match version mite_web_extras {v_mite_web_extr["version_mite_data"]}. \n"
+                f"Was mite_web_extras updated?"
+            )
+
+        os.remove(src)
+        shutil.rmtree(trgt)
+
+        logger.info(
+            "DownloadManager: Completed organization of mite_web_extras download"
+        )
+
+    def generate_summary(self) -> None:
+        """Creates summary file from metadata_general.json"""
+        summary = {"entries": {}}
+
+        with open(self.location.joinpath("metadata_general.json")) as infile:
+            metadata = json.load(infile)
+
+        for key, val in metadata["entries"].items():
+            summary["entries"][key] = {
+                "accession": val["accession"],
+                "status": val["status_icon"],
+                "status_plain": val["status"],
+                "name": val["enzyme_name"],
+                "tailoring": val["tailoring"],
+                "cofactors_organic": val["cofactors_organic"],
+                "cofactors_inorganic": val["cofactors_inorganic"],
+                "description": val["enzyme_description"],
+                "reaction_description": val["reaction_description"],
+                "organism": val["organism"],
+                "domain": val["domain"],
+                "kingdom": val["kingdom"],
+                "phylum": val["phylum"],
+                "class": val["class"],
+                "order": val["order"],
+                "family": val["family"],
+            }
+
+        keys = sorted(summary.get("entries").keys())
+        summary = {"entries": {key: summary["entries"][key] for key in keys}}
+
+        with open(self.location.joinpath("summary.json"), "w") as f:
+            f.write(json.dumps(summary))