+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] 27-10-2025
+
+### Changed
+
+- Simplified docker build - auxiliary files are downloaded
+- Reworked submission portal - show pending and submitted entries
+
 ## [1.6.6] 15-10-2025
 
 ### Changed